+++ src/grandorgue/midi/events/GOMidiEventPattern.cpp
@@ -41,10 +41,14 @@ void GOMidiEventPattern::DeviceIdFromYaml(
 }
 
 int GOMidiEventPattern::convertValueBetweenRanges(
-  int srcValue, int srcLow, int srcHigh, int dstLow, int dstHigh) {
-  const int dstAbsLow = std::min(dstLow, dstHigh);
-  const int dstAbsHigh = std::max(dstLow, dstHigh);
-  int dstValue = srcValue - srcLow;
+  int8_t srcValue,
+  int8_t srcLow,
+  int8_t srcHigh,
+  int8_t dstLow,
+  int8_t dstHigh) {
+  const int8_t dstAbsLow = std::min(dstLow, dstHigh);
+  const int8_t dstAbsHigh = std::max(dstLow, dstHigh);
+  int8_t dstValue = srcValue - srcLow;
 
   if (srcHigh != srcLow)
     dstValue = dstLow
